# api/openapi/order.yml
openapi: 3.0.3
info:
  title: Order Service
  description: Order Service
  version: 1.0.0
servers:
  - url: "https://{hostname}/api"
    variables:
      hostname:
        default: 127.0.0.1

paths:
  /customer/{customer_id}/orders/{order_id}:
    get:
      description: "Get order"
      parameters:
        - in: path
          name: customer_id
          schema:
            type: string
          required: true

        - in: path
          name: order_id
          schema:
            type: string
          required: true

      responses:
        "200":
          description: todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

        default:
          description: todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /customer/{customer_id}/orders:
    post:
      description: "Create order"
      parameters:
        - in: path
          name: customer_id
          schema:
            type: string
          required: true

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"

      responses:
        "200":
          description: todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

        default:
          description: todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /products:
    get:
      description: "Get all products information"
      responses:
        "200":
          description: A list of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        default:
          description: todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Order:
      type: object
      required:
        - id
        - customer_id
        - status
        - items
        - payment_link
      properties:
        id:
          type: string
        customer_id:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"

        payment_link:
          type: string

    Item:
      type: object
      required:
        - id
        - name
        - quantity
        - price_id
      properties:
        id:
          type: string
        name:
          type: string
        quantity:
          type: integer
          format: int32
        price_id:
          type: string

    Error:
      type: object
      properties:
        message:
          type: string

    CreateOrderRequest:
      type: object
      required:
        - customer_id
        - items
      properties:
        customer_id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithQuantity"

    ItemWithQuantity:
      type: object
      required:
        - id
        - quantity
      properties:
        id:
          type: string
        quantity:
          type: integer
          format: int32
          minimum: 1

    Product:
      type: object
      required:
        - id
        - name
        - quantity
      properties:
        id:
          type: string
        name:
          type: string
        quantity:
          type: integer
          format: int32

    ProductListResponse:
      type: object
      required:
        - errorno
        - message
        - data
        - trace_id
      properties:
        errorno:
          type: integer
        message:
          type: string
        data:
          type: object
          required:
            - products
          properties:
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
        trace_id:
          type: string

    Response:
      type: object
      properties:
        errorno:
          type: integer
        message:
          type: string
        data:
          type: object
        trace_id:
          type: string
      required:
        - errorno
        - message
        - data
        - trace_id