# internal/order/app/dto/product.go
package dto

type ProductDTO struct {
	ID       string `json:"id"`
	Name     string `json:"name"`
	Quantity int32  `json:"quantity"`
}

type GetProductsResponse struct {
	Products []ProductDTO `json:"products"`
}